Rewrote a status update method.
Previous method (using WORLD_MAP_UPDATE event) was resource-intensive due to the fact, that this event is firing dozens of times when you're standing near the border of a zone or a region. This caused huge framerate issues in places like Black Mountain, some spots in Timeless Isle, and almost all borders of zones and regions. New method uses kind of timer loop, which checks if you're in a new zone or a region every 2 seconds.