Fix decoding and blurring issue for scaled or old (3.3.5) UI (#5)
Fixes #4 and #1.

Versions of WoW before 3.3.5 had a problem with processing frames that stood close to each other, which caused blurring between them, which did not allow normal processing of a multicolored array of data (incorrect colors caused the script to crash and display "In Main Menu"). Moreover, if the user had changed in-game scaling or used scaling in the OS itself, it resulted in incorrect calculation of pixel size, which also caused blurring.

In this Pull Request, I have rewritten the algorithm for reading a multicolored array of data to account for these cases, and I have also added an automatic scale edit for the addon frames themselves to be independent of the player interface scaling. Special thanks to @CheeseKeik for helping me come up with and write the algorithm